requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appmap==2.1.7
click==8.1.7
colorama==0.4.6
fastapi==0.112.2
h11==0.14.0
idna==3.8
importlib-resources==5.13.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
PyYAML==6.0.2
slack_sdk==3.31.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
